moved pages to separate directory
